Add icons and title for mobile app display
Add tennis ball icons as favicon for general use by browsers and an apple touch icon. The latter will be used when adding the link to the home screen on iOS devices. The title of the app on the home screen is the same as the title of the web page.